+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [6.38.0](https://github.com/commercelayer/commercelayer-sdk/compare/v6.37.0...v6.38.0) (2025-04-03)
+
+
+### Features
+
+* update resources to schema v7.8.4 ([25217b2](https://github.com/commercelayer/commercelayer-sdk/commit/25217b2c943061315ea6a5e9f5b35fffac82b69f))
+
+# [6.37.0](https://github.com/commercelayer/commercelayer-sdk/compare/v6.36.0...v6.37.0) (2025-03-28)
+
+
+### Features
+
+* update resources to schema v7.8.3 ([6f4510a](https://github.com/commercelayer/commercelayer-sdk/commit/6f4510a3b2a04b90b113c38f089453332e8c995d))
+
+# [6.36.0](https://github.com/commercelayer/commercelayer-sdk/compare/v6.35.0...v6.36.0) (2025-03-18)
+
+
+### Features
+
+* update resources to schema v7.8.2 ([3e297b0](https://github.com/commercelayer/commercelayer-sdk/commit/3e297b0c3e61a2790a921561aa390fe59471e7c3))
+
 # [6.35.0](https://github.com/commercelayer/commercelayer-sdk/compare/v6.34.0...v6.35.0) (2025-03-06)